Move loading out into carto-vector-tile-loader.ts
