+++ modules/carto/src/layers/carto-tile-layer.ts
@@ -11,81 +11,18 @@ import {
   _TileLoadProps as TileLoadProps
 } from '@deck.gl/geo-layers';
 import {GeoJsonLayer} from '@deck.gl/layers';
-import {geojsonToBinary} from '@loaders.gl/gis';
-import {KeyValueProperties, Tile, TileReader} from './schema/carto-tile';
 import {TileFormat, TILE_FORMATS} from '../api/maps-api-common';
-import {LoaderOptions, LoaderWithParser} from '@loaders.gl/loader-utils';
-import type {BinaryFeatures} from '@loaders.gl/schema';
 import type {Feature} from 'geojson';
 
-function parseJSON(arrayBuffer: ArrayBuffer): any {
-  return JSON.parse(new TextDecoder().decode(arrayBuffer));
-}
+import CartoVectorTileLoader from './schema/carto-vector-tile-loader';
 
-function parsePbf(buffer: ArrayBuffer): Tile {
-  const pbf = new Protobuf(buffer);
-  const tile = TileReader.read(pbf);
-  return tile;
-}
-
-function unpackProperties(properties: KeyValueProperties[]) {
-  if (!properties || !properties.length) {
-    return [];
-  }
-  return properties.map(item => {
-    const currentRecord: Record<string, unknown> = {};
-    item.data.forEach(({key, value}) => {
-      currentRecord[key] = value;
-    });
-    return currentRecord;
-  });
-}
-
-function parseCartoTile(arrayBuffer: ArrayBuffer, options?: LoaderOptions): BinaryFeatures | null {
-  if (!arrayBuffer) return null;
-  const formatTiles = options && options.cartoTile && options.cartoTile.formatTiles;
-  if (formatTiles === TILE_FORMATS.GEOJSON) return geojsonToBinary(parseJSON(arrayBuffer).features);
-
-  const tile = parsePbf(arrayBuffer);
-
-  const {points, lines, polygons} = tile;
-  const data = {
-    points: {...points, properties: unpackProperties(points.properties)},
-    lines: {...lines, properties: unpackProperties(lines.properties)},
-    polygons: {...polygons, properties: unpackProperties(polygons.properties)}
-  };
-
-  // Note: there is slight, difference in `numericProps` type, however geojson/mvtlayer can cope with this
-  return data as unknown as BinaryFeatures;
-}
-
+// TODO do not define default twice
 const defaultTileFormat = TILE_FORMATS.BINARY;
-
-const CartoTileLoader: LoaderWithParser = {
-  name: 'CARTO Tile',
-  version: '1',
-  id: 'cartoTile',
-  module: 'carto',
-  extensions: ['pbf'],
-  mimeTypes: [
-    'application/vnd.carto-vector-tile',
-    'application/x-protobuf' // Back-compatibility
-  ],
-  category: 'geometry',
-  worker: false,
-  parse: async (arrayBuffer, options) => parseCartoTile(arrayBuffer, options),
-  parseSync: parseCartoTile,
-  options: {
-    cartoTile: {
-      formatTiles: defaultTileFormat
-    }
-  }
-};
 
 const defaultProps: DefaultProps<CartoTileLayerProps> = {
   ...MVTLayer.defaultProps,
   formatTiles: defaultTileFormat,
-  loaders: [CartoTileLoader]
+  loaders: [CartoVectorTileLoader]
 };
 
 /** All properties supported by CartoTileLayer. */
@@ -122,7 +59,7 @@ export default class CartoTileLayer<
 
     loadOptions = {
       ...loadOptions,
-      mimeType: 'application/x-protobuf'
+      mimeType: 'application/vnd.carto-vector-tile'
     };
 
     if (formatTiles) {
